add helper for pushing queries
This permits us to separate pushing on the query stack
from executing the query.
